Factor out FHR v release reconciliation
Both the chartsync package and the operator package need to install or
upgrade helm releases depending on the definitions given in
FluxHelmRelease resources. To deduplicate, drive these operations
through the chartsync, by factoring out the bit which compares a
FluxHelmRelease to the state of a release, and takes action if
necessary.

The operator package also deletes releases, when it's told a
FluxHelmRelease has been deleted. So that we only have one thing
through which we operate on the cluster, I've added that to chartsync
as well.

When starting, the operator is told about all existing FluxHelmRelease
resources as though they had been added; previously this meant that it
would (often redundantly) upgrade all the associated releases. Since
the reconcilation compares what would be released with what _is_
released, before issuing an upgrade, a side-effect of this factoring
is that the redundant upgrades aren't done.