Implement `force` for manual releases
A manual release through the API or then with fluxctl may come across
cases where configuration prevents it to perform a certain update. That
can either be a locked controller, or a container image that is
filtered. This PR adds `--force` to the CLI and `Force` to the release
spec in the API to circumvent restrictions on a per-command basis.

For locked controllers, `--force` will disregard their locked state when
controllers are listed separeately. But `--all` invocation will ignore
the `--force` flag.

For container tag filters, `--force` will allow to update to images
outside of that specification. Again, using `--update-all-images` will
remove the effect of `--force` for images.

For older daemons used in conjunction with a new fluxctl, the `--force`
param will just be ignored which should be OK since it's doing less than
intended.